feat: Refactor Cloudflare integration and improve VM management logic

+++ serve/createVm.js
@@ -7,45 +7,50 @@ const Azure = require('./MachineCreates/azure');
 
 let statusRef;
 let allowedIpaddrsRef;
+let nextDeleteTime = 0; // Sonraki silme işlemi için zaman kaydı
 
 setInterval(() => {
     let dontCloseVm = false;
     if (!statusRef) return;
-    Object.entries(statusRef).map(([ip, server]) => {
+    const cpuUsages = Object.entries(statusRef).map(([ip, server]) => {
         const lastStatus = server.status[server.status.length - 1];
-        if (lastStatus < config.CpuReleaseUsage.min) dontCloseVm = true;
+        if(!lastStatus) return 0;
+        return Number(lastStatus.cpu) || 0;
     });
-    if (dontCloseVm) return;//cpu kullanımı hala fazla ise bırakılıyor
+    const avgCpuUsage = cpuUsages.reduce((sum, usage) => sum + usage, 0) / cpuUsages.length;
+    if (avgCpuUsage < config.CpuReleaseUsage.min) dontCloseVm = true;
+    if (dontCloseVm) return; //cpu kullanımı hala fazla ise bırakılıyor
+
     const date = new Date();
-    const aServer = servers.find(z => z.time + config.CpuReleaseUsage.removeVmTimeout > date);
+    if (date.getTime() < nextDeleteTime) return; // Silme zamanı gelmediyse çık
+
+    const aServer = servers.find(z => z.time + config.CpuReleaseUsage.removeVmTimeout < date);
     if (!aServer) return;
 
-    deleteVm(aServer.rule)
+    aServer.rule.ip = aServer.serve.ip;
+    deleteVm(aServer.rule);
 
     aServer.serve.delete();
     servers.splice(servers.indexOf(aServer), 1);
+
+    nextDeleteTime = date.getTime() + config.CpuReleaseUsage.nextRemoveVmTimeout; // Yeni silme zamanı ayarla
 }, 1000)
 
 
 async function createVm(status, allowedIpaddrs) {
-    console.log("DEBUG");
-
     statusRef = status || {};
     allowedIpaddrsRef = allowedIpaddrs || [];
 
-    const availableServers = config.Machines.find(z =>
-        servers.filter(y => y.serverConfig.type == z.type).length < z.count + 1
+    const serverConfig = config.Machines.find(z =>
+        (servers.filter(y => y.serverConfig.type == z.type)?.length || 0) < z.count + 1
     );
-
-    if (!availableServers || !availableServers.length) return;
-    const serverConfig = availableServers[0];
-
+    
     switch (serverConfig.type) {
-        case 'azure':
+        case 'Azure':
 
             const serve = await Azure.create(serverConfig);
 
-            servers.push({
+            servers.push({//Burada olmasını sebebi count'u aşmaması
                 serverConfig,
                 serve,
                 time: Date.now()
@@ -53,7 +58,8 @@ async function createVm(status, allowedIpaddrs) {
 
             await serve;
             
-            aServer.serve.rule = await addVm(serve.ip, serve.resourceGroup)
+            serve.rule = await addVm(serve.ip, serve.resourceGroup)
+            serve.url = config.domain.replace("{server}", serve.rule.name);
             allowedIpaddrsRef.push(serve.ip);
             console.log(serve.url);
